refactor: 쿼리함수 분리
